Added files for new analysis and benchmarking
-compConcordIndxModded.R doesn't is just like compConcordIndx.R execpt
that it doesn't include iorio, iskar or superpred. This is because I want
to test larger benchmarks, and those other methods have been precomputed
on smaller benchmarks.

-constLuminexLayer.R creates an affinity matrix based on the pearson
correlation between drugs for the various luminex features.

-createDrugCommonsBenchmark.R subsets the drug target commons data based
on the drugs that are found in the sensitivity layer and creates a csv
file with two columns: MOLECULE_NAME, TARGET_NAME.

-createLuminexData.R subsets the luminex feature based on the drugs found
in the sensitivity layer and creates a matrix where column names are drugs
adrugTargetBenchModded.R now combines the selected drug target dataset
with more targets from DTC in the hopes of creating a larger benchmark.

-generateDrugPairsModded.R omits iorio and iskar for the sake of having a
larger benchmark.

-luminexData.r opens the subsetted luminex data file, cleans up drug
names, and returns subsets the data further based on the drugs that are
found in both perturbation data and sensitivity data.

-
a larger benchmark.

-

-