Moved KNN files to a more appropriate dir
-Added script for doing "CV" with KNN. We can't really do CV due to the
nature of the problem. The SNF algo prevents us from doing CV since if we
build a fused network on just a subset of drugs, then we can't really see
where a new drug fits into the picture. In order to classify a new
drug, it's similarities with other drugs has to be computed and that
starts at the feature level.