Added notebooks to compare new pert sigs
-There are now also notebooks that show the performance of the vanilla
model

-Changed the way stackedLayerAnalysis works so that it now accepts
arguments that can provide common drugs as well as an argument that
selects whether to use the old or new perturbation signatures.
Essentially, it is less hard coded and more dynamic.

-Added notebool to show the effect of imputing more and more perturbation
drugs in the hope of showing that the additive effect coming from the
perturbation layer is important to the integrated AUC. This doesn't seem
to be the case for the ctrpv2 drug targe benchmark.