Added notebooks for investigating performance
-A notebook for looking at the drugs that the model wasn't able to
predict targets for. Also, checks to see if different combinations of
layers make different mistakes.

-A notebook for looking at how the sensitivity of the knn method (without
limiting targets to top 5) varies as the number of neighbours is
increased. Also checks to see if there are certain drug classes that
require particularly large number of neighbours to identify a target.

-A notebook for looking at how the positions of the top predicted targets
fare in terms of correclty predicting actual targets.