Added notebooks for CCA and more KNN
-The CCA notebook is made for comparing CCA performance of the new
perturbation signatures to just standard pearson correlation.

-The knnWithUncertaintyInfo notebook shows how our nearest neighbours
method disregards neighbours that don't have drug target info available.
This explains quite a few of the misclassifications made.

-The largestPossibleNetwork notebook shows off our imputation method even
more by having a fused network the size of the sensitivity layer (669
drugs). AUC performance isn' bad, but the KNN performance for this network
is awful, which has yet to be explained.